# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  response.WhatsAppBaseResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server untuk mengelola WhatsApp messaging, status, dan webhook
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WhatsApp API Server
  version: "1.0"
paths:
...
  /whatsapp/device/qr-image:
    get:
      consumes:
      - application/json
      description: Get the QR code as PNG image for WhatsApp authentication (can be
        directly used in <img> tags)
      produces:
      - image/png
      responses:
        "200":
          description: QR Code PNG image
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get WhatsApp QR code as PNG image
      tags:
      - WhatsApp Device
...
  /whatsapp/qr:
    get:
      consumes:
      - application/json
      description: Get the QR code for WhatsApp authentication
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WhatsAppBaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get WhatsApp QR code
      tags:
      - WhatsApp
...
  /whatsapp/v1/media/download:
    get:
      consumes:
      - application/json
      description: Download and proxy WhatsApp media files (images, videos, audio,
        documents) with Persistent Local Storage
      parameters:
      - description: Session ID
        in: query
        name: session_id
        required: true
        type: string
      - description: WhatsApp media URL
        in: query
        name: url
        required: true
        type: string
      - description: Media type (image/video/audio/document)
        in: query
        name: type
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Download WhatsApp media
      tags:
      - WhatsApp Messages
...
  /whatsapp/v1/messages/stream:
    get:
      consumes:
      - application/json
      description: Get real-time message updates via SSE for a specific session
      parameters:
      - description: Session ID
        in: query
        name: session_id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Stream real-time messages
      tags:
      - WhatsApp Messages
...
schemes:
- http
- https
swagger: "2.0"